chore: move websocket manager to utils
